requirements.txt
Django==6.0.4
gunicorn==23.0.0
python-dotenv==1.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3
